+++ Block/Redirect/Pending.php
@@ -118,4 +118,5 @@ class Pending extends \Magento\Framework\View\Element\Template
         }
         return null;
     }
+
 }